refactor: DefaultProfileImageService 패키지 변경
